Modifying srtp installation and compilation methods Fixed encoder bugs that occur when encoder settings are empty

+++ src/projects/transcode/transcode_stream.h
@@ -134,6 +134,6 @@ private:
 
 	uint8_t _stats_queue_full_count;
 
-	uint8_t _max_queue_size;
+	uint64_t _max_queue_size;
 };
